separate into compute module
