+++ python/geoarrow-compute/README.md
@@ -1,0 +1,7 @@
+# `geoarrow.rust.compute`
+
+Compute operations on GeoArrow data.
+
+## Documentation
+
+Refer to the documentation at [geoarrow.org/geoarrow-rs/python](https://geoarrow.org/geoarrow-rs/python).
