+++ python/geoarrow-compute/pyproject.toml
@@ -1,0 +1,19 @@
+[build-system]
+requires = ["maturin>=1.4.0,<2.0"]
+build-backend = "maturin"
+
+[project]
+name = "geoarrow-rust-compute"
+requires-python = ">=3.8"
+dependencies = ["arro3-core", "geoarrow-rust-core", "pyproj"]
+classifiers = [
+    "Programming Language :: Rust",
+    "Programming Language :: Python :: Implementation :: CPython",
+    "Programming Language :: Python :: Implementation :: PyPy",
+]
+dynamic = ["version"]
+
+[tool.maturin]
+features = ["pyo3/extension-module"]
+module-name = "geoarrow.rust.core._compute"
+python-source = "python"
