+++ python/geoarrow-compute/python/geoarrow/rust/compute/__init__.py
@@ -1,0 +1,5 @@
+from . import enums
+from ._compute import *
+from ._compute import ___version
+
+__version__: str = ___version()
