+++ python/geoarrow-compute/src/algorithm/geo/area.rs
