+++ python/geoarrow-compute/src/algorithm/geo/centroid.rs
