+++ python/geoarrow-compute/src/algorithm/geo/chaikin_smoothing.rs
