+++ python/geoarrow-compute/src/algorithm/geo/scale.rs
