+++ python/geoarrow-compute/src/algorithm/geo/simplify.rs
