+++ python/geoarrow-compute/src/algorithm/native/explode.rs
@@ -1,8 +1,9 @@
-use crate::interop::util::{pytable_to_table, table_to_pytable};
 use geoarrow::algorithm::native::ExplodeTable;
 use pyo3::prelude::*;
 use pyo3_arrow::PyTable;
 use pyo3_geoarrow::PyGeoArrowResult;
+
+use crate::util::{pytable_to_table, table_to_pytable};
 
 #[pyfunction]
 pub fn explode(py: Python, input: PyTable) -> PyGeoArrowResult<PyObject> {
