+++ python/geoarrow-compute/src/algorithm/native/total_bounds.rs
