+++ python/geoarrow-compute/src/broadcasting/primitive.rs
