+++ python/geoarrow-compute/src/ffi/mod.rs
@@ -1,0 +1,4 @@
+//! Arrow FFI via the C Data Interface and the Arrow PyCapsule Interface.
+
+pub mod from_python;
+pub mod to_python;
