+++ python/geoarrow-compute/src/ffi/to_python/mod.rs
@@ -1,0 +1,3 @@
+pub mod array;
+
+pub use array::{chunked_geometry_array_to_pyobject, geometry_array_to_pyobject};
